requirements.txt
spotify
requests
python-dotenv